asg.backend/requirements.txt
aniso8601==9.0.1
asgiref==3.5.1
black==22.3.0
click==8.1.3
Django==4.0.4
django-filter==21.1
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.3
django-quill-editor==0.1.40
graphene==3.0
graphene-django==3.0.0b7
graphql-core==3.1.7
graphql-relay==3.1.5
install==1.3.5
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.1.0
platformdirs==2.5.2
promise==2.3
PyJWT==2.3.0
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.2.0
